Desktop Application
Developed a desktop application for Healthcare Appointment System called "HealthHarmony" using WPF for the user interface, C# for application logic, and SQLite database for local data storage. Implemented functionality to manage doctor data with create, read, update, and delete operations, integrated with a SQLite database for persistent storage. This is a simple and interactive application which allow doctors to log into the app and change or update their details at anytime effectively.